requirements.txt
pyteal==0.23.0
py-algorand-sdk==2.3.0
beaker-pyteal==1.0.0
python-dotenv==0.19.2
